requirements.txt
flask
Flask-SQLAlchemy
SQLAlchemy
flask-restful
flask-cors
pillow
matplotlib
pandas